Another bug Fix
Used cornerX/Y instead of x/y for outof bound checking on the right and bottom
Changed coordinates from uint8_t (would have breaked once we implemented emuRectFill) to int16_t (int8_t doesn't provide enought possible values to be safe imo)

+++ src/translation.hpp
@@ -17,8 +17,8 @@
 
 #define sprtSize 8
 
-extern int16_t pico8XOrgin;
-extern int16_t pico8YOrgin;
+extern int pico8XOrgin;
+extern int pico8YOrgin;
 extern bool screenShake;
 extern bool pauseEmu;
 extern void *gameState;
